cleanup and add hedgedReads along with concurrentServerCalls

+++ hbase-hadoop-compat/src/main/java/org/apache/hadoop/hbase/client/CallStats.java
@@ -56,11 +56,11 @@ public class CallStats {
     this.callTimeMs = callTimeMs;
   }
 
-  public int getConcurrentCallsPerServer() {
+  public long getConcurrentCallsPerServer() {
     return concurrentCallsPerServer;
   }
 
-  public void setConcurrentCallsPerServer(int callsPerServer) {
+  public void setConcurrentCallsPerServer(long callsPerServer) {
     this.concurrentCallsPerServer = callsPerServer;
   }
